[refactor][test] Move assert equals and retry to a base class (#14815)
(cherry picked from commit d0d9aa583b1ed1fc4387443d0d4d81c616ad77ff)
